# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive Driver Tracking SaaS Application for Indonesian Fleet Management",
        "title": "FleetTracker Pro API",
        "termsOfService": "https://fleettracker.id/terms",
        "contact": {
            "name": "FleetTracker Pro Support",
            "url": "https://fleettracker.id/support",
            "email": "support@fleettracker.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/payments/bank-transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create direct bank transfer - not implemented, use invoice generation and manual confirmation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create bank transfer payment (Not Implemented)",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/e-wallet": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create e-wallet payment - not supported, using manual bank transfer instead",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create e-wallet payment (Not Implemented)",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/invoices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve invoices for the authenticated company with filtering and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get company invoices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice status (pending, paid, overdue, cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of invoices to return (default: 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of invoices to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/payment.InvoiceResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a new invoice with Indonesian PPN 11% tax calculation for manual bank transfer payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Generate invoice for manual bank transfer",
                "parameters": [
                    {
                        "description": "Invoice generation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.InvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/payment.InvoiceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/invoices/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Confirm payment for an invoice via manual bank transfer with Indonesian banking system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Confirm manual bank transfer payment",
                "parameters": [
                    {
                        "description": "Payment confirmation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.PaymentConfirmationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/invoices/{id}/instructions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed payment instructions for manual bank transfer with Indonesian banking details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get payment instructions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/payment.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/payment.PaymentInstructions"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/qris": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create QRIS payment - not supported, using manual bank transfer instead",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create QRIS payment (Not Implemented)",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/subscriptions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get company subscriptions - not implemented yet",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get subscriptions (Not Implemented)",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new subscription - not implemented yet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create subscription (Not Implemented)",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/subscriptions/billing": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate automatic billing for company subscription with Indonesian PPN 11% tax calculation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Generate subscription billing",
                "parameters": [
                    {
                        "description": "Subscription billing data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.SubscriptionBillingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/payment.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "payment.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Validation failed"
                },
                "error": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid request data"
                }
            }
        },
        "payment.InvoiceRequest": {
            "type": "object",
            "required": [
                "billing_period",
                "company_id"
            ],
            "properties": {
                "billing_period": {
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "subscription_id": {
                    "type": "string"
                }
            }
        },
        "payment.InvoiceResponse": {
            "type": "object",
            "properties": {
                "due_date": {
                    "type": "string"
                },
                "invoice_id": {
                    "type": "string"
                },
                "invoice_number": {
                    "type": "string"
                },
                "invoice_pdf": {
                    "description": "Base64 encoded PDF",
                    "type": "string"
                },
                "payment_instructions": {
                    "$ref": "#/definitions/payment.PaymentInstructions"
                },
                "subtotal": {
                    "type": "number"
                },
                "tax_amount": {
                    "type": "number"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
        "payment.PaymentConfirmationRequest": {
            "type": "object",
            "required": [
                "bank_account",
                "invoice_id",
                "transfer_amount",
                "transfer_date"
            ],
            "properties": {
                "bank_account": {
                    "type": "string"
                },
                "invoice_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "reference_number": {
                    "type": "string"
                },
                "transfer_amount": {
                    "type": "number"
                },
                "transfer_date": {
                    "type": "string"
                }
            }
        },
        "payment.PaymentInstructions": {
            "type": "object",
            "properties": {
                "account_holder": {
                    "type": "string"
                },
                "account_number": {
                    "type": "string"
                },
                "amount": {
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "reference_code": {
                    "type": "string"
                },
                "transfer_note": {
                    "type": "string"
                }
            }
        },
        "payment.SubscriptionBillingRequest": {
            "type": "object",
            "required": [
                "billing_cycle",
                "company_id",
                "end_date",
                "start_date",
                "subscription_id"
            ],
            "properties": {
                "billing_cycle": {
                    "description": "monthly, yearly",
                    "type": "string"
                },
                "company_id": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "subscription_id": {
                    "type": "string"
                }
            }
        },
        "payment.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}